btrfs: add support for 3- and 4- copy striped profile, raid10c34
Add new incompatible feature, two new profiles similar to the RAID1C34,
the striped version. They could have been added back then in one go but
nobody suggested that and the use case has not been recognized.

On setups with many disks the RAID1C34 profiles tend to be on a few
devices, which could become noticeable as this is recommended for
metadata. The striped profiles can improved that. The main point here is
it should be used for higher number of devices, 14 and more.

The RAID10C3 could be compared to RAID6, providing the same redundancy
and striping, though the space utilization is worse. OTOH the repair
logic is much simpler and builds on existing proven code.

The RAID10C4 profile is for feature parity. The degenerate mode is
supported for both as we have that now for RAID0 and RAID10.

Signed-off-by: David Sterba <dsterba@suse.com>