btrfs: use embedded bio for super block write
The super block bio for submission is allocated and freed at each write.
Though the allocation is backed by a bioset and can't fail, it adds
some pressure to the subsystem.

We can avoid that by using an embedded bio, similar to the flush bio,
that spans the device lifetime and we can reinitialize it for each
write. The size of struct btrfs_device increases by about 112 bytes, but
the overall size is still under 1K so it fits into the same slab.

Signed-off-by: David Sterba <dsterba@suse.com>