btrfs: reduce extent tree lock contention when searching for inline backref
When inserting extent backref, in order to check whether refs other than
inline refs are used, we always use keep locks for tree search, which
will increase the lock contention of extent-tree.

We do not need the parent node every time to determine whether normal
refs are used.
It is only needed when the extent-item is the last item in a leaf.

Therefore, we change to first use keep_locks=0 for search.
If the extent-item happens to be the last item in the leaf, we then
change to keep_locks=1 for the second search to reduce lock contention.

Reviewed-by: Filipe Manana <fdmanana@suse.com>
Signed-off-by: Robbie Ko <robbieko@synology.com>
Signed-off-by: David Sterba <dsterba@suse.com>